Provide a interface for Bitmarket Protocol
The new interface `BitmarketProtocol` would represent the operations
that may be used by any peer in order to support the protocol.

In order to decouple the protocol as much as possible from its
implementation (protobuf), this interface uses its own types. Some
utility classes for converting from and to the implementation types are
provided.