Add sequence number to offers and some renaming
For offer discovery and status updating, it is useful to have a sequence
number of each offer message. Let’s say an offer is published, and a
peer accepts to assume part of it. After exchange is done, the offer
have to be published again, indicating the original amount less the part
accepted in the exchange. This may be done by propagating the offer
again with a new sequence number and the new quantity, so any peer may
know that the offered amount has changed.

This change also renames `offer.proto` to the more appropriate name
`bitmarket.proto`. The outer class `OfferProtocol` has been changed
to `BitmarketProtobuf`.