Change protobuf service interface of server
Broadcast server is serving the `PeerService` interface. So far, the
interface of both peers and broadcast server was the same, but now
the necessity to add more operations make it more useful to have
independent services for broadcast and peers.

This change adds a new service `BroadcastService` and refactors the
broadcast server to implement such an interface.