Make forward gateway strongly typed
One of the issues arised about fordware gateway is that it accepts
messages of type `Any` while forwarding. This type unsafe may derive
in problems due to the lack of type checking.

This change introduces a `MessageWrite` generic trait that is able
to write a message of type `T` using proto RPC or any other
communication mechanism we could incorporate in the future. This trait
can be implicitly implemented for each message type we want to forward.
The `ForwardMessage` actor message expects an instance of `MessageWrite`
for the message we want to forward. That means a compilation error on
any attempt to forward a unknown message object.