Turn `PaymentProcessor` into an actor
`PaymentProcessor` trait is aimed to be used in an asynchronous way by
producing `Future` objects. Since we already have Akka to deal with
async IO, it is convenient to consider the payment processors as
another actor.

This change turns `PaymentProcessor` trait into a hierarchy of
actors. The same functionality is implemented in a message-passing
approach.

This addresses [task 52](http://redmine.coinffeine.com/issues/52).