Add a wallet actor class hierarchy
The wallet actor is responsible to wrap a wallet and operate on its
funds. Under requirement, it will create a multisign transaction and
will block its founds so they are not used in any other operation.
The resulting tx can be broadcast so its funds would be spent, or
it can be released by the wallet actor under demand, turning its funds
available again.